@@ -12,17 +12,17 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^4.4.7",
-    "@angular/common": "^4.4.7",
-    "@angular/compiler": "^4.4.7",
-    "@angular/core": "^4.4.7",
-    "@angular/forms": "^4.4.7",
-    "@angular/http": "^4.4.7",
-    "@angular/platform-browser": "^4.4.7",
-    "@angular/platform-browser-dynamic": "^4.4.7",
-    "@angular/router": "^4.4.7",
-    "@ngx-translate/core": "^8.0.0",
-    "@ngx-translate/http-loader": "^2.0.1",
+    "@angular/animations": "^6.0.9",
+    "@angular/common": "^6.0.9",
+    "@angular/compiler": "^6.0.9",
+    "@angular/core": "^6.0.9",
+    "@angular/forms": "^6.0.9",
+    "@angular/http": "^6.0.9",
+    "@angular/platform-browser": "^6.0.9",
+    "@angular/platform-browser-dynamic": "^6.0.9",
+    "@angular/router": "^6.0.9",
+    "@ngx-translate/core": "^10.0.2",
+    "@ngx-translate/http-loader": "^3.0.1",
     "angular-sortablejs": "^2.5.2",
     "angular2-websocket": "^0.9.6",
     "bootstrap": "^4.1.2",
@@ -33,7 +33,8 @@
     "ngx-bootstrap": "^2.0.5",
     "ngx-dropzone-wrapper": "^4.9.0",
     "ngx-toastr": "^8.9.1",
-    "rxjs": "^5.5.11",
+    "rxjs": "^6.2.2",
+    "rxjs-compat": "^6.0.0-rc.0",
     "sortablejs": "^1.7.0",
     "tether": "^1.4.4",
     "web-animations-js": "^2.3.1",
@@ -42,8 +43,8 @@
   "devDependencies": {
     "@angular-devkit/build-angular": "~0.6.8",
     "@angular/cli": "^6.0.8",
-    "@angular/compiler-cli": "^4.4.7",
-    "@angular/language-service": "^4.4.7",
+    "@angular/compiler-cli": "^6.0.9",
+    "@angular/language-service": "^6.0.9",
     "@types/jasmine": "~2.5.53",
     "@types/jasminewd2": "^2.0.3",
     "@types/node": "^6.0.114",
@@ -60,6 +61,6 @@
     "protractor": "~5.1.2",
     "ts-node": "~3.2.0",
     "tslint": "~5.3.2",
-    "typescript": "~2.3.3"
+    "typescript": "~2.7.2"
   }
 }
