@@ -1,8 +1,8 @@
+import { HttpClient } from '@angular/common/http';
 import { Composition } from './../models/composition';
 import { Observable } from 'rxjs/Rx';
 import { Set } from './../models/set';
 import { Injectable } from '@angular/core';
-import { ApiService } from './api.service';
 import { Response } from '@angular/http';
 import { environment } from '../../environments/environment';
 import { CompositionFile } from '../models/composition-file';
@@ -10,14 +10,14 @@ import { CompositionFile } from '../models/composition-file';
 @Injectable()
 export class FileService {
 
-  constructor(private apiService: ApiService) { }
+  constructor(private http: HttpClient) { }
 
   getFiles(): Observable<CompositionFile[]> {
-    return this.apiService.get('file/list')
-      .map((response: Response) => {
+    return this.http.get('file/list')
+      .map((response: Array<Object>) => {
         let files: CompositionFile[] = [];
 
-        for (let file of response.json()) {
+        for (let file of response) {
           files.push(Composition.getFileObjectByType(file));
         }
 
@@ -26,7 +26,7 @@ export class FileService {
   }
 
   deleteFile(file: CompositionFile): Observable<void> {
-    return this.apiService.post('file/delete?name=' + file.name + '&type=' + file.type, undefined).map((response: Response) => {
+    return this.http.post('file/delete?name=' + file.name + '&type=' + file.type, undefined).map((response: Response) => {
       return null;
     });
   }
