@@ -1,37 +1,37 @@
+import { HttpClient } from '@angular/common/http';
 import { Observable } from 'rxjs/Observable';
 import { Injectable } from '@angular/core';
-import { ApiService } from './api.service';
 import { Version } from '../models/version';
 
 @Injectable()
 export class UpdateService {
 
-  constructor(private apiService: ApiService) { }
+  constructor(private http: HttpClient) { }
 
   // Get the version of the device
   getCurrentVersion(): Observable<Version> {
-    return this.apiService.get('system/current-version')
-    .map(result => {
-      return new Version(result.json());
+    return this.http.get('system/current-version')
+    .map(response => {
+      return new Version(response);
     });
   }
 
   // Get the latest available version
   getRemoteVersion(): Observable<Version> {
-    return this.apiService.get('system/remote-version')
-    .map(result => {
-      return new Version(result.json());
+    return this.http.get('system/remote-version')
+    .map(response => {
+      return new Version(response);
     });
   }
 
   doUpdate(): Observable<null> {
-    return this.apiService.post('system/update', null).map(() => {
+    return this.http.post('system/update', null).map(() => {
       return null;
     });
   }
 
   finishUpdate(): Observable<null> {
-    return this.apiService.post('session/dismiss-update-finished', null).map(() => {
+    return this.http.post('session/dismiss-update-finished', null).map(() => {
       return null;
     });
   }
